+++ src/Dosepy/tools/gamma.py
@@ -1,9 +1,9 @@
 import copy
 
 import numpy as np
-from scipy.ndimage import sobel
+from scipy.ndimage import sobel, zoom
 
-from Dosepy.image import ArrayImage
+from Dosepy.image import ArrayImage, MM_PER_INCH
 
 def is_close(val: float, target: float, delta: float):
     """Return whether the value is neat the target."""
@@ -26,92 +26,104 @@ class GammaManager:
 
         method : str
             ExhaustiveSearch: Compute gamma for every pixel.
-            Chi: .
+            Chi: Computes comparision based on `Bakai et al <http://iopscience.iop.org/0031-9155/48/21/006/>`_ eq.6,
+        params: {dose_ta: value, dist_ta: value, }
         """
+        # TODO
         if method == "ExhaustiveSearch":
-            return evaluated.gamma2D()[0]
+            return evaluated.gamma2D(
+                 dose_ta=3,
+                 dist_ta=3
+            )[0]
+        
+        if method == "FastGamma_Chi":
+             return chi(
+                  reference_image=reference,
+                  comparison_image=evaluated,
+             )
         
 
-    def chi(
-            self,
-            reference_image: ArrayImage,
-            comparison_image: ArrayImage,
-            doseTA: float = 1,
-            distTA: float = 1,
-            threshold: float = 0.1,
-            normalize: bool = True,
-        ) -> np.ndarray:
-            """Calculate the gamma between the current image (reference) and a comparison image.
-            Adapted from pylinac.
+def chi(
+        reference_image: ArrayImage,
+        comparison_image: ArrayImage,
+        doseTA: float = 3,
+        distTA: float = 3,
+        threshold: float = 10,
+        interpolate: bool = True,
+    ) -> np.ndarray:
+        """Calculate the gamma between the current image (reference) and a comparison image.
+        Adapted from pylinac.
 
-            The gamma calculation is based on `Bakai et al
-            <http://iopscience.iop.org/0031-9155/48/21/006/>`_ eq.6,
-            which is a quicker alternative to the standard Low gamma equation.
+        The gamma calculation is based on `Bakai et al
+        <http://iopscience.iop.org/0031-9155/48/21/006/>`_ eq.6,
+        which is a quicker alternative to the standard Low gamma equation.
 
-            Parameters
-            ----------
-            reference_image : {:class:`~Dosepy.image.ArrayImage`}
-                The reference image.
-            comparison_image : {:class:`~Dosepy.image.ArrayImage`}
-                The comparison image. The image must have the same DPI/DPMM to be comparable.
-                The size of the images must also be the same.
-            doseTA : int, float
-                Dose-to-agreement in percent; e.g. 2 is 2%.
-            distTA : int, float
-                Distance-to-agreement in mm.
-            threshold : float
-                The dose threshold percentage of the maximum dose, below which is not analyzed.
-                Must be between 0 and 100.
-            normalize : bool
-                Whether to normalize the images. This sets the max value of each image to the same value.
+        Parameters
+        ----------
+        reference_image : {:class:`~Dosepy.image.ArrayImage`}
+            The reference image.
+        comparison_image : {:class:`~Dosepy.image.ArrayImage`}
+            The comparison image. The image must have the same DPI/DPMM to be comparable.
+            The size of the images must also be the same.
+        doseTA : int, float
+            Dose-to-agreement in percent; e.g. 2 is 2%.
+        distTA : int, float
+            Distance-to-agreement in mm.
+        threshold : float
+            The dose threshold percentage of the maximum dose, below which is not analyzed.
+            Must be between 0 and 100.
+        interpolate : bool
+            If True the arrays are zoomed using spline interpolation to have a spatial resolution of dpmm = 1.
 
-            Returns
-            -------
-            gamma_map : numpy.ndarray
-                The calculated gamma map.
+        Returns
+        -------
+        gamma_map : numpy.ndarray
+            The calculated gamma map.
 
-            See Also
-            --------
-            :func:`~pylinac.core.image.equate_images`
-            """
-            # error checking
-            if not is_close(reference_image.dpmm, comparison_image.dpmm, delta=1):
-                raise AttributeError(
-                    f"The image resolution do not match: {reference_image.dpi:.2f} vs. {comparison_image.dpi:.2f}"
-                )
-            same_x = is_close(reference_image.shape[1], comparison_image.shape[1], delta=1.1)
-            same_y = is_close(reference_image.shape[0], comparison_image.shape[0], delta=1.1)
-            if not (same_x and same_y):
-                raise AttributeError(
-                    f"The images are not the same size: {reference_image.shape} vs. {comparison_image.shape}"
-                )
-
-            # set up reference and comparison images
-            ref_img = ArrayImage(copy.copy(reference_image.array))
-
-            if normalize:
-                ref_img.normalize()
-            comp_img = ArrayImage(copy.copy(comparison_image.array))
-
-            if normalize:
-                comp_img.normalize()
-
-            # invalidate dose values below threshold so gamma doesn't calculate over it
-            ref_img.array[ref_img < threshold/100 * np.max(ref_img)] = np.nan
-
-            # convert distance value from mm to pixels
-            distTA_pixels = reference_image.dpmm * distTA
-
-            # construct image gradient using sobel filter
-            img_x = sobel(ref_img.as_type(np.float32), 1)
-            img_y = sobel(ref_img.as_type(np.float32), 0)
-            grad_img = np.hypot(img_x, img_y)
-
-            # equation: (measurement - reference) / sqrt ( doseTA^2 + distTA^2 * image_gradient^2 )
-            subtracted_img = np.abs(comp_img - ref_img)
-            denominator = np.sqrt(
-                ((doseTA / 100.0) ** 2) + ((distTA_pixels**2) * (grad_img**2))
+        """
+        # error checking
+        if not is_close(reference_image.dpmm, comparison_image.dpmm, delta=1):
+            raise AttributeError(
+                f"The image resolution do not match: {reference_image.dpi:.2f} vs. {comparison_image.dpi:.2f}"
             )
-            gamma_map = subtracted_img / denominator
+        same_x = is_close(reference_image.shape[1], comparison_image.shape[1], delta=1.1)
+        same_y = is_close(reference_image.shape[0], comparison_image.shape[0], delta=1.1)
+        if not (same_x and same_y):
+            raise AttributeError(
+                f"The images are not the same size: {reference_image.shape} vs. {comparison_image.shape}"
+            )
 
-            return gamma_map
+        # interpolate reference and comparison images to have a 
+        # spatial resolution of dpmm = 1 (dpi = 25.4)
+        ref_img = ArrayImage(
+            zoom(reference_image.array, zoom = 1/reference_image.dpmm),
+            dpi=MM_PER_INCH,
+        )
+
+        comp_img = ArrayImage(
+            zoom(comparison_image.array, zoom = 1/comparison_image.dpmm),
+            dpi=MM_PER_INCH,
+        )
+
+        # invalidate dose values below threshold so gamma doesn't calculate over it
+        ref_img.array[ref_img < threshold/100 * np.max(ref_img)] = np.nan
+
+        # convert distance value from mm to pixels
+        distTA_pixels = reference_image.dpmm * distTA
+
+        # Calculate doseTa
+        doseTA_Gray = doseTA/100 * np.amax(ref_img)
+
+        # construct image gradient using sobel filter
+        img_x = sobel(ref_img.as_type(np.float32), 1)
+        img_y = sobel(ref_img.as_type(np.float32), 0)
+        grad_img = np.hypot(img_x, img_y)
+
+        # equation: (measurement - reference) / sqrt ( doseTA^2 + distTA^2 * image_gradient^2 )
+        subtracted_img = np.abs(comp_img - ref_img)
+        denominator = np.sqrt(
+            ((doseTA_Gray) ** 2) + ((distTA_pixels**2) * (grad_img**2))
+        )
+        gamma_map = subtracted_img / denominator
+
+        return gamma_map